Reservation, Seat 상태 상수 위치 변경, ReservationBookingResponse 의 성공일떄 반환 객체 생성 메서드 생성
