+++ src/main/java/personnel/jupitorsendsme/pulseticket/entity/Reservation.java
@@ -1,5 +1,6 @@
 package personnel.jupitorsendsme.pulseticket.entity;
 
+import java.time.Duration;
 import java.time.LocalDateTime;
 
 import jakarta.persistence.Column;
@@ -17,7 +18,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import personnel.jupitorsendsme.pulseticket.constants.ReservationConstants;
 
 /**
  * 예약 정보를 관리하는 엔티티
@@ -29,6 +29,51 @@ import personnel.jupitorsendsme.pulseticket.constants.ReservationConstants;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Reservation extends BaseEntity {
+
+	/**
+	 * 예약 만료 시간
+	 * 임시로 24시간으로 설정
+	 */
+	public static final Duration RESERVATION_EXPIRATION = Duration.ofHours(24);
+
+	/**
+	 * Reservation 테이블의 status 컬럼에 해당하는 상태 <br>
+	 * Pending : 예약 직후 상태. 아직 결제는 안한 상태 <br>
+	 * CONFIRMED : 결제해서 예약된 상태 <br>
+	 * CANCELLED : 예약이 취소된 상태 <br>
+	 * Expired : 예약은 했으나 일정시간 결제를 안해서 만료된 상태.
+	 */
+	public enum ReservationStatus {
+
+		PENDING,
+		CONFIRMED,
+		CANCELLED,
+		EXPIRED;
+
+		public ReservationStatus confirm() {
+			if (this != PENDING) {
+				throw new IllegalStateException("예약 대기가 아닌 상태에서 예약 확정 불가, 현재 상태 : " + this);
+			}
+			return CONFIRMED;
+		}
+
+		public ReservationStatus cancel() {
+			switch (this) {
+				case CANCELLED -> throw new IllegalStateException("취소된 예약");
+				case EXPIRED -> throw new IllegalStateException("만료된 예약");
+			}
+			return CANCELLED;
+		}
+
+		public ReservationStatus expire() {
+			switch (this) {
+				case CONFIRMED -> throw new IllegalStateException("이미 결제된 예약");
+				case EXPIRED -> throw new IllegalStateException("만료된 예약");
+			}
+			return EXPIRED;
+		}
+	}
+
 	/**
 	 * 예약 고유 식별자
 	 */
@@ -62,7 +107,7 @@ public class Reservation extends BaseEntity {
 	 */
 	@Enumerated(EnumType.STRING)
 	@Column(nullable = false, length = 20)
-	private ReservationConstants.ReservationStatus status;
+	private ReservationStatus status;
 
 	/**
 	 * 예약 만료 일시
