+++ src/main/java/personnel/jupitorsendsme/pulseticket/repository/SeatRepository.java
@@ -5,7 +5,6 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import personnel.jupitorsendsme.pulseticket.constants.ReservationConstants;
 import personnel.jupitorsendsme.pulseticket.entity.Seat;
 
 /**
@@ -14,13 +13,13 @@ import personnel.jupitorsendsme.pulseticket.entity.Seat;
  */
 public interface SeatRepository extends JpaRepository<Seat, Long> {
 
-	boolean existsByEvent_IdAndStatus(Long event_id, ReservationConstants.SeatStatus status);
+	boolean existsByEvent_IdAndStatus(Long event_id, Seat.SeatStatus status);
 
 	boolean existsSeatByEvent_IdAndSeatNumberAndStatus(Long event_id, Integer seatNumber,
-		ReservationConstants.SeatStatus status);
+		Seat.SeatStatus status);
 
 	Optional<Seat> findByEvent_IdAndSeatNumberAndStatus(Long event_id, Integer seatNumber,
-		ReservationConstants.SeatStatus status);
+		Seat.SeatStatus status);
 
 	List<Seat> findByEvent_Id(Long event_id);
 }
